+++ country.py
@@ -1,6 +1,1 @@
 # start_value / gdp
-count   ry_st = {'Russia': [700, 1.09],
-           'USA': [400, 1.2],
-           'German': [550, 1.12],
-           'China': [300, 1.25],
-           'Sweden': [500, 1.17]}
